fix(issies/4): https://github.com/8xFF/atm0s-media-server-console/issues/4
