+++ apps/console/next.config.mjs
@@ -1,6 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  // output: 'export',
+  output: 'export',
   transpilePackages: ['@packages/ui', '@packages/core', '@packages/react-hooks', '@packages/react-ui'],
 }
