+++ apps/console/src/app/zone/detail/components/gateways.tsx
