+++ apps/console/src/app/zone/detail/components/index.ts
