+++ apps/console/src/app/zone/detail/components/medias.tsx
