+++ apps/console/src/app/zone/detail/containers/index.ts
@@ -1,0 +1,1 @@
+export * from './zone-detail'
