+++ apps/console/src/app/zone/detail/page.tsx
@@ -1,0 +1,5 @@
+import { ZoneDetail } from './containers'
+
+export default function ZoneDetailScreen() {
+  return <ZoneDetail />
+}
