+++ apps/console/src/app/zone/list/components/zone-item.tsx
@@ -12,7 +12,7 @@ type Props = {
 export const ZoneItem: React.FC<Props> = ({ zone }) => {
   const router = useRouter()
   return (
-    <Card className="cursor-pointer" onClick={() => router.push(`/zones/${zone?.zone_id}`)}>
+    <Card className="cursor-pointer" onClick={() => router.push(`/zone/detail?zone_id=${zone?.zone_id}`)}>
       <CardHeader className="flex flex-row items-center justify-between space-y-0 p-3 pb-0">
         <CardTitle className="text-sm font-medium">zone_id: {zone?.zone_id}</CardTitle>
         <SignalIcon className="text-emerald-500" size={16} />
