+++ apps/console/src/app/zone/list/containers/index.ts
@@ -1,0 +1,1 @@
+export * from './zone-list'
