+++ apps/console/src/app/zone/list/page.tsx
@@ -1,0 +1,5 @@
+import { ZoneList } from './containers'
+
+export default function ZonesScreen() {
+  return <ZoneList />
+}
